Split charts MCP tool registration
